chore: organize files structure
