+++ src/components/common/layout.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
-import "../styles/global-theme.css"
-import "../styles/global-styles.css"
+import "../../styles/global-theme.css"
+import "../../styles/global-styles.css"
 import Header from "./header"
 interface LayoutProps {
   children?: React.ReactNode
