--- src/components/fallback.tsx
@@ -1,8 +1,0 @@
-interface FallbackProps {
-  blocktype: string
-}
-
-export default function Fallback({ blocktype }: FallbackProps) {
-  console.warn(`No component found for: ${blocktype}`)
-  return null
-}
